fixed some, works well

+++ backend/app/routers/upload.py
@@ -5,6 +5,7 @@ from app.database import get_session
 from app.models.models import File as FileModel
 import shutil
 import os
+import hashlib
 
 router = APIRouter()
 
@@ -37,16 +38,26 @@ def upload_and_save_file(
     file: UploadFile = File(...),
     session: Session = Depends(get_session)
 ):
+    content= file.file.read()
+    file_hash = hashlib.sha256(content).hexdigest()
+    existing_file = session.query(FileModel).filter_by(hash=file_hash).first()
+    if existing_file: 
+        return {
+            "message": "File already exists. ", 
+            "url": existing_file.url, 
+            "filename": existing_file.filename
+        }
     file_path = os.path.join(UPLOAD_DIR, file.filename)
     with open(file_path, "wb") as buffer:
         shutil.copyfileobj(file.file, buffer)
 
     file_record = FileModel(
         filename=file.filename,
-        url=f"/uploads/{file.filename}"
+        url=f"/uploads/{file.filename}", 
+        hash = file_hash
     )
     session.add(file_record)
     session.commit()
     session.refresh(file_record)
 
-    return {"url": file_record.url}
+    return {"url": file_record.url, "filename": file_record.filename}
